+++ packages/market/src/test/MarketFacade.test.ts
@@ -59,11 +59,9 @@ describe('Market-Facade', () => {
     const issuerAccount = web3.eth.accounts.privateKeyToAccount(issuerPK).address;
 
     it('should set ERC20 token', async () => {
-        erc20TestTokenAddress = (await deployERC20TestToken(
-            web3,
-            accountTrader,
-            privateKeyDeployment
-        )).contractAddress;
+        erc20TestTokenAddress = (
+            await deployERC20TestToken(web3, accountTrader, privateKeyDeployment)
+        ).contractAddress;
 
         erc20TestToken = new Erc20TestToken(web3, erc20TestTokenAddress);
     });