+++ packages/market/src/utils/migrateContracts.ts
@@ -11,12 +11,16 @@ export async function migrateMarketRegistryContracts(
 ): Promise<MarketLogic> {
     const privateKeyDeployment = deployKey.startsWith('0x') ? deployKey : `0x${deployKey}`;
 
-    const marketLogicAddress = (await deploy(web3, MarketLogicJSON.bytecode, {
-        privateKey: privateKeyDeployment
-    })).contractAddress;
+    const marketLogicAddress = (
+        await deploy(web3, MarketLogicJSON.bytecode, {
+            privateKey: privateKeyDeployment
+        })
+    ).contractAddress;
 
     const marketLogic = new MarketLogic(web3, marketLogicAddress);
-    await marketLogic.initialize(certificateLogicAddress);
+    await marketLogic.initialize(certificateLogicAddress, {
+        privateKey: privateKeyDeployment
+    });
 
     return marketLogic;
 }